+++ spec/system/vendor_api/vendor_receives_application_spec.rb
@@ -63,6 +63,7 @@ RSpec.feature 'Vendor receives the application', recruitment_cycle: 2020 do
           nationality: %w[GB US],
           domicile: @application.domicile,
           uk_residency_status: 'UK Citizen',
+          uk_residency_status_code: 'A',
           fee_payer: '02',
           english_main_language: true,
           other_languages: nil,